+++ client/Packages/com.beamable.server/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `IUsageApi` is available
+
 ### Fixed
 
 - Microservice docker logs no longer put quotes around every log parameter